+++ client/lib/models/state/game/posessions/assets/cash_asset_item.dart
@@ -12,6 +12,5 @@ abstract class CashAssetItem
   CashAssetItem._();
 
   String get name;
-
-  int get value;
+  double get value;
 }
